Fixing variable type errors

+++ pipelining_2.c
@@ -12,7 +12,7 @@ This code contains software pipelining for both y and x of SAD calculation
 #define BITS_PER_PIXEL_OFFSET 0x001C
 #define DATA_OFFSET_OFFSET 0x000A
 
-void readImage(char filename[], uint32_t pixels[HEIGHT][WIDTH])
+void readImage(char filename[], uint8_t pixels[HEIGHT][WIDTH])
 {
     FILE *bmp = fopen(filename, "rb");
     // find pixel data offset
@@ -40,7 +40,7 @@ void readImage(char filename[], uint32_t pixels[HEIGHT][WIDTH])
     fclose(bmp);
 }
 
-int calc_block_diff(int x_first_pixel, int y_first_pixel, int x_second_pixel, int y_second_pixel, uint32_t image_first[HEIGHT][WIDTH], uint32_t image_second[HEIGHT][WIDTH])
+int calc_block_diff(int x_first_pixel, int y_first_pixel, int x_second_pixel, int y_second_pixel, uint8_t image_first[HEIGHT][WIDTH], uint8_t image_second[HEIGHT][WIDTH])
 {
     int SAD_temp = 0;
     int temp_image_first = image_first[y_first_pixel][x_first_pixel];
